requirements.txt
pydantic-settings>=2.0.0
tenacity>=8.2.0
prometheus-client>=0.20.0
langchain-gigachat>=0.3.0
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=2.0.0
langchain-core>=0.3.0
langchain-community>=0.3.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-telegram-bot>=21.0
httpx>=0.27.0
duckduckgo-search>=7.0.0
tabulate>=0.9.0
statsmodels>=0.14.0
plotly>=5.0.0
kaleido>=0.2.1